Add workflow-inspector with saved filed view (#3)
This is a re-implementation of
https://github.com/haesleinhuepf/napari-workflow-inspector

This version maintains the same matplotlib style task graph inspector.
It has the addition of uploading a `workflow.yaml` to inspect, which
previous versions didn't. It's npe2 compatible and modern napari ready.
It has not been validated for its work yet with live workflow production
(like with napari-assistant), but that code path is hopefully in tact.

Because this adds no dependencies, I thought it fine to just live in
this repo, but it could be refactored out in the future.

This was generated mostly with the help of copilot, and perhaps I will
regret adding this because I have little interest in actively
maintaining qt code, instead preferring magicgui widgets and such. But,
I do think this is most helpful for collaborators, and is a sanity check
that my new workflow format is set up correctly.

The plot won't generate without matplotlib, but other parts of the
inspector work, so I added it as an optional to keep dependencies
slimmer for now. _Most_ environments seem to end up with matplotlib
eventually so I'm not too worried.